+++ server/db/models/secret.spec.js
@@ -1,51 +1,18 @@
-/* global describe beforeEach it */
-
 const { expect } = require('chai');
 const db = require('../index');
 const Secret = db.model('secret');
-const User = db.model('user');
-const { seedSecrets, seedUsers } = require('./test-seeds');
-const assert = require('assert');
 
 describe('Secret model', () => {
-  before(async () => {
-    await db.sync({ force: true });
-    const users = await User.bulkCreate(seedUsers);
-    const secrets = await Secret.bulkCreate(seedSecrets);
-    assert(users.length === 3, 'User.bulkCreate failed');
-    assert(secrets.length === 3, 'Secret.bulkCreate failed');
+  describe('properties', () => {
+    it('should include secret message', () => {
+      expect(Secret.attributes.message).to.be.an('object');
+      expect(Secret.attributes.message.allowNull).to.equal(false);
+      expect(Secret.attributes.message.notEmpty).to.equal(true);
+    });
+
+    it('should include a boolean for isPublic', () => {
+      expect(Secret.attributes.isPublic).to.be.an('object');
+      expect(Secret.attributes.isPublic.defaultValue).to.equal(false);
+    });
   });
-
-  it('should be true', () => {
-    expect(true).to.equal(true);
-  });
-
-  // beforeEach(() => {
-  //   return Passage.create({
-  //     title: 'Hello World',
-  //     content: 'Hello yee world of mine!'
-  //   })
-  //   .then(passage => {
-  //     hello = passage;
-  //   })
-  //   .catch(err => {
-  //     console.log(err.message);
-  //   });
-  // });
-
-  // describe('Model assignments', () => {
-  //   it('Defaults to not public', () => {
-  //     expect(hello.isPublic).to.be.equal(false);
-  //   });
-  // });
-
-
-  // describe('Methods', () => {
-
-  //   it('Preview returns a slice the length you provide it', () => {
-  //     expect(hello.preview(3).length).to.be.equal(3);
-  //     expect(hello.preview(3)).to.be.equal('Hel');
-  //   });
-  // });//end instance methods
-}); // end describe('Passage model')
-
+});
